+++ DataModels/Entities/EmailToken.cs
@@ -4,11 +4,11 @@ namespace DataModels.Entities
 {
     public class EmailToken
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Token { get; set; }
         public DateTime? CreatedOn { get; set; }
         public DateTime? ExpireOn { get; set; }
-        public int? UserId { get; set; }
+        public long? UserId { get; set; }
         public string EmailType { get; set; }
     }
 }
